fix dd logging tests
